pyproject.toml
[project]
name = "scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dacite>=1.9.2",
    "psycopg[binary,pool]>=3.2.9",
    "tornado>=6.5.1",
    "yoyo-migrations>=9.0.0",
    "zmq>=0.0.0",
]

[dependency-groups]
dev = [
    "pyhamcrest>=2.1.0",
    "pytest>=8.4.1",
    "testcontainers[postgres]>=4.12.0",
]

